Full edit and copying of model data into final report
Added my model code to the final report sheet. Also went through and continued to copy over work from the planning sheet, and do edit passes on content to try and make it look more like a cohesive report. Added some notes of my own to the Discussion section based on the findings of the model.